Convert townName
